Fix delay in dispatch_after_short Update provision profiles

+++ Swift-Base/Helpers/FSHelpers.swift
@@ -114,6 +114,6 @@ extension Double {
     }
 }
 
-func dispatch_after_short (delay:CGFloat, block:dispatch_block_t) {    
-    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (Int64(delay) * Int64(NSEC_PER_SEC))), dispatch_get_main_queue(), block);
+func dispatch_after_short (delay:Double, block:dispatch_block_t) {
+    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, Int64(delay * Double(NSEC_PER_SEC))), dispatch_get_main_queue(), block);
 }
